Pull out a `goto_page` method
There's a lot of duplication in the URL constants. Also, I'm planning to
move the URLs into a common base class, and it's a pain to reference
namespaced constants.

This refactor creates a frozen hash of page => URL, along with a method
to visit a URL with a given name.