Extract a find_receiver method
This extraction is a little more challenging. Not only do we to perform
a search on the website, but we need some branching logic based on the
contents of the page we get back. This extraction hides that complexity
by reducing the page state to a small set of symbols representing the
actual states that we care about. The calling function is able to branch
appropriately based on the returned symbols.