Extract SetReceivers object
The object we've been working with is called SetReceiver (singular), but
the code is looping over a list of receiver objects. This was handy for
testing, but it's confusing and easy to fix.

This extraction pulls out a separate object to handle instantiating and
looping through the list. As a bonus, we can move the current receiver
into an instance variable so we don't have to pass it around between
the interal methods.