Make call a single level of abstraction
The call method reads pretty cleanly, but it's structurally quite nasty.
It's a mix of logical process instructions, low-level browser instructions
and a nasty conditional switching on magic symbol values.

This refactoring makes call very clean with a single level of abstraction.
It moves all of the complexity into an existing conditional, and moves
the explicit page navigation into the method that actually cares about
that navigation.