Make search_for_receiver take a block
The last refactor has one glaring problem: implicit shared state.
The search_for_receiver method (non obviously) sets some state
(the contents of the current page) that every other method call
depends on. It would be very easy for an unrelated change to
search_for_receiver to unknowingly break find_receiver.

We can't easily get rid of the requirement for shared state, but
we can make the presence of the shared state more explicit. By
making search_for_receiver take a block it's more obvious that
methods are executing in the context created by that method.